📦 chore(Dockerfile): add Dockerfile to project for containerization
The Dockerfile is added to the project to enable containerization of the application. It uses the `python:3.9-slim` base image and sets the working directory to `/app`. The `pyproject.toml` and `poetry.lock` files are copied to the working directory and Poetry is installed along with the project dependencies. Finally, the entire project is copied to the container and the command `python main.py` is set as the default command to run when the container starts.